PS Angular 13.9 - Feature Modules
We created a feature module--product.module.ts-- to hold the imports that can run the application. After creating the module file through the Angular CLI, we moved many of the Components there from the AppModule. The terminal command to generate this file and import AppModule was:
ng g m products/product --flat -m app